+++ pornolab-enhancer/features/image-view/image-url-extractor.js
@@ -24,6 +24,9 @@ export default (function () {
   }
 
   const extractors = [
+    // link:      http://fastpic.ru/view/90/2017/0206/1f4072e748ac53657de2056ef8498adb.png.html
+    // thumbnail: http://i90.fastpic.ru/thumb/2017/0206/db/1f4072e748ac53657de2056ef8498adb.jpeg
+    // image:     http://i90.fastpic.ru/big/2017/0206/db/1f4072e748ac53657de2056ef8498adb.png?noht=1
     {
       name: 'FastPic',
       allowed: true,
@@ -41,6 +44,9 @@ export default (function () {
       }
     },
 
+    // link:      http://img45.imagevenue.com/img.php?image=10934_bscap0004_122_1068lo.jpg
+    // thumbnail: http://img45.imagevenue.com/loc1068/th_10934_bscap0004_122_1068lo.jpg
+    // image:     http://img45.imagevenue.com/aAfkjfp01fo1i-3071/loc1068/10934_bscap0004_122_1068lo.jpg
     {
       name: 'ImageVenue',
       allowed: true,
@@ -60,6 +66,9 @@ export default (function () {
       }
     },
 
+    // link:      https://www.turboimagehost.com/p/36160085/1.jpg.html
+    // thumbnail: https://s7d8.turboimg.net/t1/36160085_1.jpg
+    // image:     https://s7d8.turboimg.net/sp/a5ce6305fca205a76c6b23a7f33262ec/1.jpg
     {
       name: 'TurboImageHost',
       allowed: true,
@@ -71,6 +80,9 @@ export default (function () {
 
     // not allowed below
 
+    // link:      http://www.imagebam.com/image/4a9c52356295333
+    // thumbnail: http://thumbnails112.imagebam.com/35630/4a9c52356295333.jpg
+    // image:     http://112.imagebam.com/download/FllgT6YgLpm0PUEDQ9ISag/35630/356295333/%20%282%29.jpg
     {
       name: 'ImageBam',
       linkSelector: '[href^="http://www.imagebam.com/image"]',
@@ -79,6 +91,9 @@ export default (function () {
       getUrl: getUrlFromPage
     },
 
+    // link:      http://imagetwist.com/awxdf3ez7gdg/cc_Barbara_Zatler__pbsi_PM200909__04XL.jpg.html
+    // thumbnail: http://img8.imagetwist.com/th/01318/awxdf3ez7gdg.jpg
+    // image:     http://img8.imagetwist.com/i/01318/awxdf3ez7gdg.jpg/cc_Barbara_Zatler__pbsi_PM200909__04XL.jpg
     {
       name: 'ImageTwist',
       linkSelector: '[href^="http://imagetwist.com"]',
@@ -86,6 +101,7 @@ export default (function () {
 
       async getUrl (extractor, link) {
         const imageName = link.href.split('/').pop()
+          .replace('.html', '')
         const extension = imageName.split('.').pop()
         const imageUrl = getThumbnailUrl(link)
           .replace('/th/', '/i/')
@@ -94,7 +110,7 @@ export default (function () {
         return `${imageUrl}${extension}/${imageName}`
       }
     },
-
+    // Remove ?
     {
       name: 'PicShick',
       linkSelector: '[href^="http://picshick.com"]',
@@ -109,7 +125,7 @@ export default (function () {
         return `${imageUrl}/${imageName}`
       }
     },
-
+    // Remove ?
     {
       name: 'imgbum',
       linkSelector: '[href^="http://imgbum.net"]',
@@ -120,6 +136,9 @@ export default (function () {
       }
     },
 
+    // link:      http://picforall.ru/348/206637/
+    // thumbnail: http://picforall.ru/allimage/10/206637-thumb.jpeg
+    // image:     http://p0xpicmoney.ru/allimage/10/206637.jpeg
     {
       name: 'PicForAll',
       linkSelector: '[href^="http://picforall.ru"]',
@@ -132,6 +151,9 @@ export default (function () {
       }
     },
 
+    // link:      http://picage.ru/10887/5456787/
+    // thumbnail: http://s4.picage.ru/y2016/01-24/10887/5456787-thumb.jpeg
+    // image:     http://s4.pic4you.ru/y2016/01-24/10887/5456787.jpeg
     {
       name: 'picage',
       linkSelector: '[href^="http://picage.ru"]',
@@ -143,7 +165,7 @@ export default (function () {
           .replace('-thumb', '')
       }
     },
-
+    // Remove ?
     {
       name: 'PixSense',
       linkSelector: '[href^="http://www.pixsense.net"]',
@@ -155,7 +177,7 @@ export default (function () {
           .replace('/small/', '/big/')
       }
     },
-
+    // Remove ?
     {
       name: 'nikapic.ru',
       linkSelector: '[href^="http://nikapic.ru"]',
@@ -166,7 +188,7 @@ export default (function () {
           .replace('/small/', '/big/')
       }
     },
-
+    // Remove ?
     {
       name: 'imgtaxi.com',
       linkSelector: '[href^="https://imgtaxi.com"]',
@@ -175,6 +197,32 @@ export default (function () {
       async getUrl (extractor, link) {
         return getThumbnailUrl(link)
           .replace('/small/', '/big/')
+      }
+    },
+
+    // link:      http://imgbox.com/Hn0YZvqQ
+    // thumbnail: http://t.imgbox.com/Hn0YZvqQ.jpg
+    // image:     https://images3.imgbox.com/56/b6/Hn0YZvqQ_o.jpg
+    {
+      name: 'imgbox.com',
+      linkSelector: '[href^="http://imgbox.com"]',
+      linkRegEx: new RegExp('^http://imgbox.com'),
+      imageUrlRegEx: /href="([^"]*)".*icon-cloud-download/,
+      getUrl: getUrlFromPage
+    },
+
+    // link:      http://payforpic.ru/49/253590/
+    // thumbnail: http://payforpic.ru/allimage/3/253590-thumb.jpeg
+    // image:     http://picker-click.ru/allimage/3/253590.jpeg
+    {
+      name: 'payforpic.ru',
+      linkSelector: '[href^="http://payforpic.ru"]',
+      linkRegEx: new RegExp('^http://payforpic.ru'),
+
+      async getUrl (extractor, link) {
+        return getThumbnailUrl(link)
+          .replace('payforpic', 'picker-click')
+          .replace('-thumb', '')
       }
     }
   ]